@@ -14,21 +14,18 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
-        "facade/ignition-contracts": "^1.0",
-        "filp/whoops": "^2.14.3",
-        "symfony/console": "^5.0"
+        "php": "^8.0.0",
+        "facade/ignition-contracts": "^1.0.2",
+        "filp/whoops": "^2.14.5",
+        "symfony/console": "^6.0.2"
     },
     "require-dev": {
-        "brianium/paratest": "^6.1",
-        "fideloper/proxy": "^4.4.1",
-        "fruitcake/laravel-cors": "^2.0.3",
-        "laravel/framework": "8.x-dev",
-        "nunomaduro/larastan": "^0.6.2",
-        "nunomaduro/mock-final-classes": "^1.0",
-        "orchestra/testbench": "^6.0",
-        "phpstan/phpstan": "^0.12.64",
-        "phpunit/phpunit": "^9.5.0"
+        "brianium/paratest": "^6.4.1",
+        "laravel/framework": "^9.0",
+        "nunomaduro/larastan": "^1.0.2",
+        "nunomaduro/mock-final-classes": "^1.1.0",
+        "orchestra/testbench": "^7.0.0",
+        "phpunit/phpunit": "^9.5.11"
     },
     "autoload-dev": {
         "psr-4": {
@@ -50,6 +47,9 @@
         "sort-packages": true
     },
     "extra": {
+        "branch-alias": {
+            "dev-develop": "6.x-dev"
+        },
         "laravel": {
             "providers": [
                 "NunoMaduro\\Collision\\Adapters\\Laravel\\CollisionServiceProvider"
