@@ -13,7 +13,6 @@ namespace Asm89\Stack;
 
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 
 class Cors implements HttpKernelInterface
 {
@@ -27,17 +26,17 @@ class Cors implements HttpKernelInterface
      */
     private $cors;
 
-    private $defaultOptions = array(
-        'allowedHeaders'         => array(),
-        'allowedMethods'         => array(),
-        'allowedOrigins'         => array(),
-        'allowedOriginsPatterns' => array(),
-        'exposedHeaders'         => array(),
+    private $defaultOptions = [
+        'allowedHeaders'         => [],
+        'allowedMethods'         => [],
+        'allowedOrigins'         => [],
+        'allowedOriginsPatterns' => [],
+        'exposedHeaders'         => [],
         'maxAge'                 => 0,
         'supportsCredentials'    => false,
-    );
+    ];
 
-    public function __construct(HttpKernelInterface $app, array $options = array())
+    public function __construct(HttpKernelInterface $app, array $options = [])
     {
         $this->app  = $app;
         $this->cors = new CorsService(array_merge($this->defaultOptions, $options));
