@@ -18,22 +18,22 @@ class CorsService
 {
     private $options;
 
-    public function __construct(array $options = array())
+    public function __construct(array $options = [])
     {
         $this->options = $this->normalizeOptions($options);
     }
 
-    private function normalizeOptions(array $options = array())/*: array*/
+    private function normalizeOptions(array $options = [])/*: array*/
     {
-        $options += array(
-            'allowedOrigins' => array(),
-            'allowedOriginsPatterns' => array(),
+        $options += [
+            'allowedOrigins' => [],
+            'allowedOriginsPatterns' => [],
             'supportsCredentials' => false,
-            'allowedHeaders' => array(),
-            'exposedHeaders' => array(),
-            'allowedMethods' => array(),
+            'allowedHeaders' => [],
+            'exposedHeaders' => [],
+            'allowedMethods' => [],
             'maxAge' => 0,
-        );
+        ];
 
         // normalize array('*') to true
         if (in_array('*', $options['allowedOrigins'])) {
@@ -145,8 +145,8 @@ class CorsService
             // Single origins can be safely set
             $response->headers->set('Access-Control-Allow-Origin', array_values($this->options['allowedOrigins'])[0]);
         } else {
-            // For dynamic headers, check the origin first
-            if ($request->headers->has('Origin') && $this->isOriginAllowed($request)) {
+            // For dynamic headers, set the requested Origin header when set and allowed
+            if ($this->isCorsRequest($request) && $this->isOriginAllowed($request)) {
                 $response->headers->set('Access-Control-Allow-Origin', $request->headers->get('Origin'));
             }
 
