@@ -3,7 +3,7 @@ import React from 'react';
 export default function Label({ forInput, value, className, children }) {
     return (
         <label htmlFor={forInput} className={`block font-medium text-sm text-gray-700 ` + className}>
-            {value ? value : { children }}
+            {value ? value : children}
         </label>
     );
 }
