@@ -13,9 +13,7 @@ class PasswordConfirmationTest extends TestCase
 
     public function test_confirm_password_screen_can_be_rendered()
     {
-        $user = Features::hasTeamFeatures()
-                        ? User::factory()->withPersonalTeam()->create()
-                        : User::factory()->create();
+        $user = User::factory()->withPersonalTeam()->create();
 
         $response = $this->actingAs($user)->get('/user/confirm-password');
 
