@@ -5,6 +5,8 @@ namespace Laravel\Telescope\Watchers;
 use Illuminate\Mail\Events\MessageSent;
 use Laravel\Telescope\IncomingEntry;
 use Laravel\Telescope\Telescope;
+use Symfony\Component\Mime\Address;
+use Symfony\Component\Mime\Part\AbstractPart;
 
 class MailWatcher extends Watcher
 {
@@ -31,16 +33,18 @@ class MailWatcher extends Watcher
             return;
         }
 
+        $body = $event->message->getBody();
+
         Telescope::recordMail(IncomingEntry::make([
             'mailable' => $this->getMailable($event),
             'queued' => $this->getQueuedStatus($event),
-            'from' => $event->message->getFrom(),
-            'replyTo' => $event->message->getReplyTo(),
-            'to' => $event->message->getTo(),
-            'cc' => $event->message->getCc(),
-            'bcc' => $event->message->getBcc(),
+            'from' => $this->formatAddresses($event->message->getFrom()),
+            'replyTo' => $this->formatAddresses($event->message->getReplyTo()),
+            'to' => $this->formatAddresses($event->message->getTo()),
+            'cc' => $this->formatAddresses($event->message->getCc()),
+            'bcc' => $this->formatAddresses($event->message->getBcc()),
             'subject' => $event->message->getSubject(),
-            'html' => $event->message->getBody(),
+            'html' => $body instanceof AbstractPart ? $body->bodyToString() : $body,
             'raw' => $event->message->toString(),
         ])->tags($this->tags($event->message, $event->data)));
     }
@@ -73,6 +77,27 @@ class MailWatcher extends Watcher
         }
 
         return isset($event->data['__telescope_queued']) ? $event->data['__telescope_queued'] : false;
+    }
+
+    /**
+     * Convert the given addresses into a readable format.
+     *
+     * @param  array|null  $addresses
+     * @return array|null
+     */
+    protected function formatAddresses(/*?*/array $addresses = null)
+    {
+        if (is_null($addresses)) {
+            return null;
+        }
+
+        return collect($addresses)->flatMap(function ($address, $key) {
+            if ($address instanceof Address) {
+                return [$address->getAddress() => $address->getName()];
+            }
+
+            return [$key => $address];
+        })->all();
     }
 
     /**
