@@ -57,9 +57,10 @@ class CollisionServiceProvider extends ServiceProvider
 
             $this->app->bind(ProviderContract::class, function () {
                 if ($this->app->has(\Facade\IgnitionContracts\SolutionProviderRepository::class)) {
-                    $solutionsRepository = new IgnitionSolutionsRepository(
-                        $this->app->get(\Facade\IgnitionContracts\SolutionProviderRepository::class)
-                    );
+                    /** @var \Facade\IgnitionContracts\SolutionProviderRepository $solutionProviderRepository */
+                    $solutionProviderRepository = $this->app->get(\Facade\IgnitionContracts\SolutionProviderRepository::class);
+
+                    $solutionsRepository = new IgnitionSolutionsRepository($solutionProviderRepository);
                 } else {
                     $solutionsRepository = new NullSolutionsRepository();
                 }
@@ -70,6 +71,7 @@ class CollisionServiceProvider extends ServiceProvider
                 return new Provider(null, $handler);
             });
 
+            /** @var \Illuminate\Contracts\Debug\ExceptionHandler $appExceptionHandler */
             $appExceptionHandler = $this->app->make(ExceptionHandlerContract::class);
 
             $this->app->singleton(
