@@ -30,7 +30,7 @@ class TentativeTypes
             'format' => 'string',
             'getTimezone' => 'DateTimeZone|false',
             'getOffset' => 'int',
-            'getTimestamp' => 'int|false',
+            'getTimestamp' => 'int',
             'diff' => 'DateInterval',
             '__wakeup' => 'void',
         ],
@@ -254,6 +254,7 @@ class TentativeTypes
             'isEquivalentTo' => 'bool',
             'isLenient' => 'bool',
             'isWeekend' => 'bool',
+            'roll' => 'bool',
             'isSet' => 'bool',
             'setTime' => 'bool',
             'setTimeZone' => 'bool',
