@@ -1,5 +1,7 @@
 <?php
 
-class ValueError extends Error
-{
+if (\PHP_VERSION_ID < 80000) {
+    class ValueError extends Error
+    {
+    }
 }
