@@ -1,9 +1,11 @@
 <?php
 
-#[Attribute(Attribute::TARGET_METHOD)]
-final class ReturnTypeWillChange
-{
-    public function __construct()
+if (\PHP_VERSION_ID < 80100) {
+    #[Attribute(Attribute::TARGET_METHOD)]
+    final class ReturnTypeWillChange
     {
+        public function __construct()
+        {
+        }
     }
 }
